moved color sorting to scripts
