+++ tests/e2e.test.js
@@ -1,4 +1,4 @@
-const { sortByColors } = require('../public/scripts/utils.js');
+const { sortByColors } = require('../scripts/color-sorting.js');
 
 const fs = require('fs');
 const path = require('path');
