+++ webpack.config.js
@@ -6,10 +6,8 @@ const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const path = require('path');
 const simpleIcons = require('simple-icons');
 
-const {
-  normalizeSearchTerm,
-  sortByColors,
-} = require('./public/scripts/utils.js');
+const { normalizeSearchTerm } = require('./public/scripts/utils.js');
+const { sortByColors } = require('./scripts/color-sorting.js');
 
 const icons = Object.values(simpleIcons);
 const sortedHexes = sortByColors(
